backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
pyjwt==2.8.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# OpenAI
openai==1.12.0

# HTTP client for Wikipedia API
httpx==0.26.0

# Environment
python-dotenv==1.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0
aiosqlite==0.19.0

# CORS
